@@ -5,15 +5,15 @@
   "dependencies": {
     "dotenv": "^8.2.0",
     "emailjs-com": "^2.4.0",
-    "firebase": "^7.1.0",
-    "node-sass": "^4.12.0",
+    "firebase": "^7.3.0",
+    "node-sass": "^4.13.0",
     "p5": "^0.9.0",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
+    "react": "^16.11.0",
+    "react-dom": "^16.11.0",
     "react-firebase-file-uploader": "^2.4.3",
     "react-measure": "^2.3.0",
     "react-p5-wrapper": "^2.0.0",
-    "react-redux": "^7.1.1",
+    "react-redux": "^7.1.3",
     "react-redux-firebase": "^2.4.1",
     "react-router-dom": "^5.1.2",
     "react-scripts": "3.0.1",
@@ -46,6 +46,6 @@
     ]
   },
   "devDependencies": {
-    "prettier": "^1.18.2"
+    "prettier": "^1.19.1"
   }
 }
