@@ -5,61 +5,70 @@
  Current features:
  1. Load audio file.
  2. Play/Pause audio.
-
- TODO: - Change volume
- TODO: - Get song duration
- TODO: - Get current progress
- TODO: - Allow users to control the part of the song to be played
+ 3. Download visualization
+ 4. Change volume
+ 5. Control the part of the song to be played
 
  ************************************************************/
 
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import { ReactComponent as SnapshotIcon } from '../../assets/PlayerBarAssets/snapshot-icon-white-dark.svg';
 import { ReactComponent as VolumeIcon } from '../../assets/PlayerBarAssets/volume-icon.svg';
 import { ReactComponent as PlayIcon } from '../../assets/PlayerBarAssets/play-icon.svg';
+import { ReactComponent as PauseIcon } from '../../assets/PlayerBarAssets/pause-icon.svg';
 import { ReactComponent as SongIcon } from '../../assets/PlayerBarAssets/song-icon.svg';
+import { ReactComponent as RollingIcon } from '../../assets/LoadingAssets/Rolling.svg';
 import UploadButton from './UploadButton/UploadButton';
 import classes from './PlayerBar.module.scss';
 import DownloadButton from './DownloadButton/DownloadButton';
 
-const pauseButton = (
-    <svg
-        xmlns="http://www.w3.org/2000/svg"
-        width="18"
-        height="24"
-        viewBox="0 0 18 24"
-    >
-        <path
-            fill="rgba(255, 255, 255, 0.75)"
-            fillRule="evenodd"
-            d="M0 0h6v24H0zM12 0h6v24h-6z"
-            className="pause-icon"
-        />
-    </svg>
-);
-
-/***********************************************************
- Parameters (src/pages/App/App.js):
- state:
- 1. volume
- 2. isPlaying
- 3. uploadedSong
- 4. isSongLoaded
-
- functions:
- 1.  onPlayPress
- *************************************************************/
-
 export default function PlayerBar(props) {
+    const [formatedDuration, setFormatedDuration] = useState('0:00');
+    const [formatedCurrentTime, setFormatedCurrentTime] = useState('0:00');
+
+    // Props from (src/pages/App/App.js):
     const {
         uploadedSong,
         duration,
         onPlayPress,
+        playPressed,
         isPlaying,
-        isSongLoaded,
+        volume,
         onVolumeChange,
-        songEnded
+        songEnded,
+        currentTime,
+        onCueTimeChange
     } = props;
+
+    // Set formated duration
+    useEffect(() => {
+        setFormatedDuration(getTime(duration));
+    }, [duration]);
+
+    // Set formated current time
+    useEffect(() => {
+        setFormatedCurrentTime(getTime(currentTime));
+    }, [currentTime]);
+
+    // Format time (eg. 140 to 2:20)
+    const getTime = dur => {
+        return (
+            Math.floor(dur / 60) + ':' + ('0' + Math.floor(dur % 60)).slice(-2)
+        );
+    };
+
+    // Update the width for the progress slider base on the current time
+    const sliderProgressWidth = {
+        width: `${(100 * currentTime) / duration}%`
+    };
+
+    // Change play button to loading svg when loading on p5 sound
+    let actionButton;
+    if (isPlaying) {
+        actionButton = <PauseIcon />;
+    } else {
+        actionButton = <RollingIcon />;
+    }
 
     return (
         <div className={classes.playerBar}>
@@ -73,16 +82,29 @@ export default function PlayerBar(props) {
             </div>
             <div className={classes.playerControls}>
                 <div className={classes.playButton} onClick={onPlayPress}>
-                    {isPlaying && isSongLoaded ? pauseButton : <PlayIcon />}
+                    {playPressed ? actionButton : <PlayIcon />}
                 </div>
                 <div className={classes.controls}>
-                    <span className={classes.progressTime}>0:00</span>
+                    <span className={classes.progressTime}>
+                        {formatedCurrentTime}
+                    </span>
                     <div className={classes.slider}>
-                        <div className={classes.progress} />
+                        <div
+                            className={classes.progress}
+                            style={sliderProgressWidth}
+                        />
+                        <input
+                            className={classes.rangeInput}
+                            type="range"
+                            min="0"
+                            max={duration}
+                            step="1"
+                            name="cue"
+                            onClick={onCueTimeChange}
+                        />
                     </div>
                     <span className={classes.progressTime}>
-                        {/* 0:00{' '} */}
-                        {uploadedSong ? duration : '0:00'}
+                        {formatedDuration}
                     </span>
                 </div>
                 <div className={classes.volume}>
@@ -91,9 +113,10 @@ export default function PlayerBar(props) {
                         <input
                             type="range"
                             min="0"
-                            max="10"
-                            step="0.5"
+                            max="1"
+                            step="0.1"
                             name="volume"
+                            value={volume}
                             className={classes.volumeSlider}
                             onChange={onVolumeChange}
                         ></input>
