@@ -1,7 +1,7 @@
 import React from 'react';
 
 const Span = ({ content, ...otherProps }) => (
-	<span {...otherProps}>{content}</span>
+    <span {...otherProps}>{content}</span>
 );
 
 export default Span;
