@@ -13,15 +13,15 @@ import './globalScss/index.scss';
 require('dotenv').config();
 
 const ROUTER = (
-	<Provider store={store}>
-		<Router>
-			<AppRouter />
-		</Router>
-	</Provider>
+    <Provider store={store}>
+        <Router>
+            <AppRouter />
+        </Router>
+    </Provider>
 );
 
 store.firebaseAuthIsReady.then(() => {
-	ReactDOM.render(ROUTER, document.getElementById('root'));
+    ReactDOM.render(ROUTER, document.getElementById('root'));
 });
 
 // If you want your app to work offline and load faster, you can change
