@@ -29,15 +29,18 @@ const [linkedIn, github] = icons;
 
 const SectionThreeIcons = () => (
     <div className={classes.socialIcons}>
-
         <ParallaxLayer
             offset={2.54}
             speed={0.5}
             id={classes.github}
             style={{ width: 'min-content', height: '0', zIndex: '99' }}
         >
-            <a href = "https://github.com/zero-to-mastery/visual-music" target="_blank">
-            <img src={github} alt="github" className={classes.socialIcon} />
+            <a
+                href="https://github.com/zero-to-mastery/visual-music"
+                target="_blank"
+                rel="noopener noreferrer"
+            >
+                <img src={github} alt="github" className={classes.socialIcon} />
             </a>
         </ParallaxLayer>
 
@@ -47,11 +50,18 @@ const SectionThreeIcons = () => (
             id={classes.linkedIn}
             style={{ width: 'min-content', height: '0', zIndex: '99' }}
         >
-            <a href = "https://www.linkedin.com/company/zerotomastery/" target="_blank">
-            <img src={linkedIn} alt="linkedIn" className={classes.socialIcon} />
+            <a
+                href="https://www.linkedin.com/company/zerotomastery/"
+                target="_blank"
+                rel="noopener noreferrer"
+            >
+                <img
+                    src={linkedIn}
+                    alt="linkedIn"
+                    className={classes.socialIcon}
+                />
             </a>
         </ParallaxLayer>
-
     </div>
 );
 
