@@ -24,7 +24,10 @@ function Login() {
         event.preventDefault();
         dispatch(cleanError());
         setSpan(
-            <img alt="loading" src={require('../../assets/loading.svg')} />
+            <img
+                alt="loading"
+                src={require('../../assets/LoadingAssets/LoadingBars.svg')}
+            />
         );
         dispatch(logIn({ email, password }));
     };
