@@ -1,7 +1,40 @@
+/*
+ Main component for Login.
+
+ Allows users to input email address and password and click
+ Login button to submit the form
+
+ Current features:
+ 1. Input user email address
+ 2. Input user password
+ 3. Do not allow email address and password filed to be blank
+ 4. Validate email address and password field.
+*/
 import React from 'react';
 import classes from './Login.module.scss';
 import { Link } from 'react-router-dom';
 import Span from '../../../components/units/Span/Span';
+import FormInput from '../../../components/units/FormInput/formInput.component';
+
+const emailValidators = [
+    {
+        validate: value => {
+            return (
+                value.includes('@') && (value.match(/@/g) || []).length === 1
+            );
+        },
+        failMsg: 'Email is not valid'
+    }
+];
+
+const passwordValidators = [
+    {
+        validate: value => {
+            return value.length >= 6 ? true : false;
+        },
+        failMsg: 'Password is too short'
+    }
+];
 
 function LoginPage({ setEmail, setPassword, onFormSubmit, span }) {
     return (
@@ -13,29 +46,21 @@ function LoginPage({ setEmail, setPassword, onFormSubmit, span }) {
                     </div>
                     <div className={classes.emailGroup}>
                         <div className={classes.emailGroup}>
-                            <label className={classes.emailLabel}>
-                                Email address
-                            </label>
-                            <input
-                                className={classes.emailField}
+                            <FormInput
+                                labelText="Email address"
                                 type="email"
-                                onChange={e => {
-                                    setEmail(e.target.value);
-                                }}
+                                validators={emailValidators}
+                                onChange={e => setEmail(e.target.value)}
                                 required
                             />
                         </div>
                     </div>
                     <div className={classes.passwordGroup}>
-                        <label className={classes.passwordLabel}>
-                            Password
-                        </label>
-                        <input
-                            className={classes.passwordField}
+                        <FormInput
+                            labelText="Password"
                             type="password"
-                            onChange={e => {
-                                setPassword(e.target.value);
-                            }}
+                            validators={passwordValidators}
+                            onChange={e => setPassword(e.target.value)}
                             required
                         />
                     </div>
